messaging_app/requirements.txt
Django==5.2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-filter==24.2
PyMySQL==1.1.1
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
flake8==7.1.0
coverage==7.5.3
mysqlclient==2.2.4

